perf(m3): sample expert weight in native dtype, drop full-tensor upcast
note_scale upcast the entire smooth-layer weight to fp32 before sampling
(`weight.detach().float()`), a ~150MB O(numel) bf16->fp32 copy per expert
mapping (128 experts x up/down x 57 MoE layers) of which only 4096 values are
kept. Gather in the native dtype first, then cast the 4096-element sample --
bit-identical result (bf16->fp32 is exact), O(count) instead of O(numel),
mirroring deterministic_sketch. The gather itself was never the cost; a 4096
index_select does not scale with tensor size.

+++ pipeline/m3_guarded_full.py
@@ -612,12 +612,15 @@ class FullGuardController:
         layer = _layer_from_name(mapping.smooth_name)
         if layer is None:
             return
-        weight = mapping.smooth_layer.weight.detach().float()
+        # Do NOT upcast the whole weight: for a MoE expert that is a 37.7M-element
+        # bf16->fp32 copy of which we keep only 4096 values. Gather in the native
+        # dtype (O(count)) and cast the tiny sample, like deterministic_sketch.
+        weight = mapping.smooth_layer.weight.detach()
         scale = scales.detach().float().to(weight.device)
         flat = weight.reshape(-1)
         count = min(DEFAULT_SKETCH_VALUES, flat.numel())
         indices = _evenly_spaced_indices(flat.numel(), count, flat.device)
-        before = flat.index_select(0, indices)
+        before = flat.index_select(0, indices).float()
         if weight.ndim == 1:
             applied_scale = scale.index_select(0, indices)
         else:
